Dynamic borders
Changes to the nature of the provinces.js file, which now creates a pol object which contains a series of keys attached to GeoJSON feature collections, or references to another key in provinces.js. The key is the year associated with the borders. The appropriate changes were made to scry.html so that this data is handled correctly.