Refactoring & overdue abstraction (Got rid of my unnecessary Graph data structure - *terrible* idea - in the front end)
Also started structuring a proper implemenation of the linking between agents.

Before, I tried a terrible and convoluted graph data structure in the front end (months ago). This was a terrible idea because, in fact, it is all about adding an extra tab, that should be part of the content property of the historical agents. The original implenation of the front end would work like a charm without any additions or major modifications if such approach is pursued.

I guess that the idea of the graph data structure is, in a way, not necessary. An adjacency list is nothing more than the same thing as adding such extra tab, to the content property of agents, listing the linked elements. Some methods to retrieve such particular information easily may or may not be necessary in the future; but, at any case, I believe such data should remain where I am pointing to.

I expect that the Agents.html will be left as it was originally (pretty much without any of my code). Good abstraction will allow us to work well without messing up the other's work so easily and unecessarily, haha. :laughing:

I'm also looking forward towards cleaning a lot of the unnecessary code I wrote while I was playing and getting familiar with the project back in Seoul.

My hope for the future is to focus primarily on the back end, which was largely missing when I started contributing - also need to refactor one or two things (today I learned about the concept of 'CallBakHell', for which I recommend this link https://medium.com/codebuddies/getting-to-know-asynchronous-javascript-callbacks-promises-and-async-await-17e0673281ee).